+++ pg/payment/member.go
@@ -1,4 +1,4 @@
-package creditcard
+package payment
 
 import (
 	"github.com/abyssparanoia/go-gmo/internal/pkg/validate"
@@ -37,4 +37,35 @@ func (cli *Client) SaveMember(
 	return res, nil
 }
 
-type UpdateMemberRequest struct{}
+// UpdateMemberRequest ... update member request
+type UpdateMemberRequest struct {
+	MemberID   string `json:"MemberID" validate:"required,max=60"`
+	MemberName string `json:"MemberName"`
+}
+
+// Validate ... validate
+func (r *UpdateMemberRequest) Validate() error {
+	return validate.Struct(r)
+}
+
+// UpdateMemberResponse ... update member response
+type UpdateMemberResponse struct {
+	MemberID string `json:"MemberID"`
+	ErrCode  string `json:"ErrCode"`
+	ErrInfo  string `json:"ErrInfo"`
+}
+
+// UpdateMember ... update member
+func (cli *Client) UpdateMember(
+	req *UpdateMemberRequest,
+) (*UpdateMemberResponse, error) {
+	if err := req.Validate(); err != nil {
+		return nil, err
+	}
+	res := &UpdateMemberResponse{}
+	_, err := cli.do(updateMemberPath, req, res)
+	if err != nil {
+		return nil, err
+	}
+	return res, nil
+}
